refactor: Move files from lib to utils folder
