+++ app/charts/area/areas-state.tsx
@@ -45,9 +45,9 @@ import {
   useTimeFormatUnit,
 } from "@/configurator/components/ui-helpers";
 import { Observation } from "@/domain/data";
-import { sortByIndex } from "@/lib/array";
-import { estimateTextWidth } from "@/lib/estimate-text-width";
 import { useLocale } from "@/locales/use-locale";
+import { sortByIndex } from "@/utils/array";
+import { estimateTextWidth } from "@/utils/estimate-text-width";
 import { makeOrdinalDimensionSorter } from "@/utils/sorting-values";
 
 export interface AreasState {
