+++ app/charts/shared/use-width.tsx
@@ -1,6 +1,6 @@
 import { createContext, ReactNode, useContext } from "react";
 
-import { useResizeObserver } from "@/lib/use-resize-observer";
+import { useResizeObserver } from "@/utils/use-resize-observer";
 
 export interface Margins {
   top: number;
