+++ app/components/chart-preview.tsx
@@ -22,8 +22,8 @@ import { ChartConfig, DataSource, useConfiguratorState } from "@/configurator";
 import { DataSetTable } from "@/configurator/components/datatable";
 import { useDataCubeMetadataQuery } from "@/graphql/query-hooks";
 import { DataCubePublicationStatus } from "@/graphql/resolver-types";
-import { useResizeObserver } from "@/lib/use-resize-observer";
 import { useLocale } from "@/locales/use-locale";
+import { useResizeObserver } from "@/utils/use-resize-observer";
 
 export const ChartPreview = ({
   dataSetIri,
