+++ app/configurator/components/chart-configurator.tsx
@@ -70,8 +70,8 @@ import {
 } from "@/graphql/query-hooks";
 import { DataCubeMetadata } from "@/graphql/types";
 import { Icon } from "@/icons";
-import useEvent from "@/lib/use-event";
 import { useLocale } from "@/locales/use-locale";
+import useEvent from "@/utils/use-event";
 
 import { ChartTypeSelector } from "./chart-type-selector";
 import useHierarchyParents from "./use-hierarchy-parents";
