+++ app/utils/array.ts
