+++ app/utils/create-chart-id.ts
