+++ app/utils/nprogress.css
